specific-scrapers/mohe-scraper/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mohe-scraper"
version = "1.0.0"
description = "Scraper for Malaysia Ministry of Higher Education (MOHE) government documents and announcements"
requires-python = ">=3.11"
authors = [{name = "Data Team"}]

dependencies = [
    "requests>=2.31.0,<3.0",
    "beautifulsoup4>=4.12.0,<5.0",
    "lxml>=4.9.0",
    "python-dateutil>=2.8.2,<3.0",
    "google-cloud-storage>=2.10.0,<3.0",
    "tenacity>=8.2.3,<9.0",
    "pydantic>=2.5.0,<3.0",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0",
    "pytest-cov>=4.1.0,<5.0",
    "black>=23.11.0,<24.0",
    "ruff>=0.1.8,<1.0",
    "mypy>=1.7.0,<2.0",
]

[project.scripts]
mohe-scraper = "mohe_scraper.cli:main"
